Loop over pairs of abundant numbers instead
Reduced the search space by instead of checking if each number can be
made from a pair of abundant numbers by checking each abundant number
less than it and it's compliment, generate all numbers that are the sum
of a pair of the known abundant numbers. Then sum all numbers not in
that set.